wip: various refactors
