+++ packages/ponder/src/graphql/types.ts
@@ -1,6 +1,6 @@
-import type { Knex } from "knex";
+import type { Database } from "better-sqlite3";
 
 type Source = { request: unknown };
-type Context = { db: Knex<Record<string, unknown>, unknown[]> };
+type Context = { db: Database };
 
 export type { Context, Source };
